+++ framework-test/src/main/java/org/checkerframework/framework/test/TestUtilities.java
@@ -38,23 +38,23 @@ import org.plumelib.util.SystemPlume;
 public class TestUtilities {
 
   /** True if the JVM is version 9 or above. */
-  public static final boolean IS_AT_LEAST_9_JVM = SystemUtil.getJreVersion() >= 9;
+  public static final boolean IS_AT_LEAST_9_JVM = SystemUtil.jreVersion >= 9;
   /** True if the JVM is version 11 or above. */
-  public static final boolean IS_AT_LEAST_11_JVM = SystemUtil.getJreVersion() >= 11;
+  public static final boolean IS_AT_LEAST_11_JVM = SystemUtil.jreVersion >= 11;
   /** True if the JVM is version 11 or lower. */
-  public static final boolean IS_AT_MOST_11_JVM = SystemUtil.getJreVersion() <= 11;
+  public static final boolean IS_AT_MOST_11_JVM = SystemUtil.jreVersion <= 11;
   /** True if the JVM is version 14 or above. */
-  public static final boolean IS_AT_LEAST_14_JVM = SystemUtil.getJreVersion() >= 14;
+  public static final boolean IS_AT_LEAST_14_JVM = SystemUtil.jreVersion >= 14;
   /** True if the JVM is version 14 or lower. */
-  public static final boolean IS_AT_MOST_14_JVM = SystemUtil.getJreVersion() <= 14;
+  public static final boolean IS_AT_MOST_14_JVM = SystemUtil.jreVersion <= 14;
   /** True if the JVM is version 16 or above. */
-  public static final boolean IS_AT_LEAST_16_JVM = SystemUtil.getJreVersion() >= 16;
+  public static final boolean IS_AT_LEAST_16_JVM = SystemUtil.jreVersion >= 16;
   /** True if the JVM is version 16 or lower. */
-  public static final boolean IS_AT_MOST_16_JVM = SystemUtil.getJreVersion() <= 16;
+  public static final boolean IS_AT_MOST_16_JVM = SystemUtil.jreVersion <= 16;
   /** True if the JVM is version 17 or above. */
-  public static final boolean IS_AT_LEAST_17_JVM = SystemUtil.getJreVersion() >= 17;
+  public static final boolean IS_AT_LEAST_17_JVM = SystemUtil.jreVersion >= 17;
   /** True if the JVM is version 17 or lower. */
-  public static final boolean IS_AT_MOST_17_JVM = SystemUtil.getJreVersion() <= 17;
+  public static final boolean IS_AT_MOST_17_JVM = SystemUtil.jreVersion <= 17;
 
   static {
     JavaCompiler compiler = ToolProvider.getSystemJavaCompiler();